requirements.txt
dashscope==1.20.4
gradio==5.4.0
python-dotenv==1.0.1
